+++ src/api/routes/followers/index.js
@@ -1,1 +1,1 @@
-import "./follower";
+import "./resonators";
